+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.5.15] - 2020-11-11
+
+- fix onError callback
+- guard onDispose with ref.mounted
+
 ## [0.5.14] - 2020-11-11
 
 - fix faulty data_state